Implement Resolve for Agent operator
(cherry picked from commit 49558165b9178b6c1bda833a48f7bfe1468c942a)
