+++ pkg/prometheus/operator_test.go
@@ -1,0 +1,41 @@
+package prometheus
+
+import "testing"
+
+func TestStatefulSetKeyToPrometheusKey(t *testing.T) {
+	cases := []struct {
+		input         string
+		expectedKey   string
+		expectedMatch bool
+	}{
+		{
+			input:         "namespace/prometheus-test",
+			expectedKey:   "namespace/test",
+			expectedMatch: true,
+		},
+		{
+			input:         "namespace/prometheus-test-shard-1",
+			expectedKey:   "namespace/test",
+			expectedMatch: true,
+		},
+		{
+			input:         "allns-z-thanosrulercreatedeletecluster-qcwdmj-0/thanos-ruler-test",
+			expectedKey:   "",
+			expectedMatch: false,
+		},
+	}
+
+	for _, c := range cases {
+		match, key := StatefulSetKeyToPrometheusKey(c.input)
+		if c.expectedKey != key {
+			t.Fatalf("Expected prometheus key %q got %q", c.expectedKey, key)
+		}
+		if c.expectedMatch != match {
+			notExp := ""
+			if !c.expectedMatch {
+				notExp = "not "
+			}
+			t.Fatalf("Expected input %sto be matching a prometheus key, but did not", notExp)
+		}
+	}
+}
