+++ pkg/prometheus/server/operator.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	"fmt"
 	"reflect"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -997,7 +996,7 @@ func (c *Operator) Resolve(ss *appsv1.StatefulSet) metav1.Object {
 		return nil
 	}
 
-	match, promKey := statefulSetKeyToPrometheusKey(key)
+	match, promKey := prompkg.StatefulSetKeyToPrometheusKey(key)
 	if !match {
 		level.Debug(c.logger).Log("msg", "StatefulSet key did not match a Prometheus key format", "key", key)
 		return nil
@@ -1021,25 +1020,6 @@ func statefulSetNameFromPrometheusName(name string, shard int) string {
 		return fmt.Sprintf("prometheus-%s", name)
 	}
 	return fmt.Sprintf("prometheus-%s-shard-%d", name, shard)
-}
-
-var prometheusKeyInShardStatefulSet = regexp.MustCompile("^(.+)/prometheus-(.+)-shard-[1-9][0-9]*$")
-var prometheusKeyInStatefulSet = regexp.MustCompile("^(.+)/prometheus-(.+)$")
-
-func statefulSetKeyToPrometheusKey(key string) (bool, string) {
-	r := prometheusKeyInStatefulSet
-	if prometheusKeyInShardStatefulSet.MatchString(key) {
-		r = prometheusKeyInShardStatefulSet
-	}
-
-	matches := r.FindAllStringSubmatch(key, 2)
-	if len(matches) != 1 {
-		return false, ""
-	}
-	if len(matches[0]) != 3 {
-		return false, ""
-	}
-	return true, matches[0][1] + "/" + matches[0][2]
 }
 
 func prometheusKeyToStatefulSetKey(key string, shard int) string {
